backend/upload_files/requirements.txt
deta
fastapi
uvicorn[standard]
email-validator
python-jose[cryptography]
gunicorn
flower
python-multipart
Jinja2
requests
xlsxwriter
Pillow